% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{add_prefix}
\alias{add_prefix}
\title{Add prefix to each element of matrix}
\usage{
add_prefix(x, prefix = "")
}
\arguments{
\item{x}{Numeric or symbolic matrix}
}
\description{
Add prefix to each element of matrix
}
\examples{
if (has_sympy()) {
  X <- matrix_sym(2, 3)
  X
  add_prefix(X, "e")
  
  X <- matrix(1:6, 3, 2)
  X
  add_prefix(X, "e")
}

}
\concept{linalg}
