% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{all_vars}
\alias{all_vars}
\title{All variables}
\usage{
all_vars(x)
}
\arguments{
\item{x}{caracas symbol}
}
\description{
Return all variables in caracas symbol
}
\examples{
if (has_sympy()){
  x <- vector_sym(5)
  all_vars(x)
}

}
\concept{caracas_symbol}
