% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{apart}
\alias{apart}
\title{Partial fraction decomposition on a rational function}
\usage{
apart(x)
}
\arguments{
\item{x}{A \code{caracas_symbol}}
}
\description{
apart() performs a partial fraction decomposition on a rational function
}
\examples{
if (has_sympy()){
  def_sym(x)
  expr = (4*x**3 + 21*x**2 + 10*x + 12)/(x**4 + 5*x**3 + 5*x**2 + 4*x)
  apart(expr)
}
}
\concept{simplify}
