% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_python.R
\name{as.expression.caracas_symbol}
\alias{as.expression.caracas_symbol}
\title{Convert caracas object to R}
\usage{
\method{as.expression}{caracas_symbol}(x, ...)
}
\arguments{
\item{x}{caracas_symbol}
}
\description{
Potentially calls \code{\link[=doit]{doit()}}.
}
\examples{
if (has_sympy()) {
  v <- vector_sym(2)
  x <- as_expr(v)
  x
  y <- as.expression(v)
  y
}

}
