% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_symbol.R
\name{as_character}
\alias{as_character}
\title{Coerce symbol to character}
\usage{
as_character(x)
}
\arguments{
\item{x}{caracas symbol}
}
\description{
Coerce symbol to character
}
\concept{caracas_symbol}
