% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{as_character_matrix}
\alias{as_character_matrix}
\title{Get matrix as character matrix}
\usage{
as_character_matrix(x)
}
\arguments{
\item{x}{caracas symbol}
}
\description{
Get matrix as character matrix
}
\concept{linalg}
