% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{as_diag}
\alias{as_diag}
\title{Construct diagonal matrix from vector}
\usage{
as_diag(x)
}
\arguments{
\item{x}{Matrix with 1 row or 1 column that is the
diagonal in a new diagonal matrix}
}
\description{
Construct diagonal matrix from vector
}
\examples{
if (has_sympy()) {
  d <- as_sym(c("a", "b", "c"))
  D <- as_diag(d)
  D
}

}
\concept{linalg}
