% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{as_expr2}
\alias{as_expr2}
\title{Convert caracas object to R (col or row wise)}
\usage{
as_expr2(x, first_doit = TRUE, column_major = TRUE)
}
\arguments{
\item{x}{caracas_symbol}

\item{first_doit}{Try \code{doit()} first}

\item{column_major}{}
}
\description{
Convert caracas object to R (col or row wise)
}
