% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{as_factor_list}
\alias{as_factor_list}
\title{Create list of factors as in a product}
\usage{
as_factor_list(...)
}
\arguments{
\item{...}{factors}
}
\description{
Create list of factors as in a product
}
\examples{
if (has_sympy()) {
 d <- 2
 m <- matrix_sym(d, d)
 mi <- inv(m)
 det_m <- det(m)
 fl <- as_factor_list(1/det_m, det_m * mi)
 tex(fl)
 m <- matrix(1:4, nrow=2)
 mi <- solve(m)
 det_m <- det(m)
 fl <- as_factor_list(1 / as_sym(det_m), det_m * mi)
 tex(fl)
}
}
