% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{as_vec}
\alias{as_vec}
\title{Stacks matrix to vector}
\usage{
as_vec(x)
}
\arguments{
\item{x}{Matrix}
}
\description{
Stacks matrix to vector
}
\examples{
if (has_sympy()) {
  A <- as_sym(matrix(1:9, 3))
  as_vec(A)
}

}
\concept{linalg}
