% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{ask}
\alias{ask}
\title{Ask for a symbol's property}
\usage{
ask(x, property)
}
\arguments{
\item{x}{symbol}

\item{property}{property, e.g. 'positive'}
}
\description{
Ask for a symbol's property
}
\concept{assumptions}
