% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_caracas_sorenh.R
\name{bdiag_}
\alias{bdiag_}
\title{Creates block diagonal caracas matrix}
\usage{
bdiag_(...)
}
\arguments{
\item{...}{Elements to be put on diagonals}
}
\value{
a caracas matrix
}
\description{
Creates block diagonal caracas matrix
}
\author{
Søren Højsgaard
}
