% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{collect}
\alias{collect}
\title{Collects common powers of a term in an expression}
\usage{
collect(x, a)
}
\arguments{
\item{x, a}{A \code{caracas_symbol}}
}
\description{
Collects common powers of a term in an expression
}
\examples{
if (has_sympy()){
  def_sym(x, y, z)
  expr = x*y + x - 3 + 2*x**2 - z*x**2 + x**3
  collect(expr, x)
}
}
\concept{simplify}
