% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{colspan}
\alias{colspan}
\title{Column space (range) of a symbolic matrix}
\usage{
colspan(x)
}
\arguments{
\item{x}{Symbolic matrix}
}
\description{
Column space (range) of a symbolic matrix
}
\examples{
if (has_sympy()) {
  X1 <- matrix_(paste0("x_",c(1,1,1,1, 2,2,2,2, 3,4,3,4)), nrow = 4)
  colspan(X1)
  X2 <- matrix_(paste0("x_",c(1,1,1,1, 0,0,2,2, 3,4,3,4)), nrow = 4)
  colspan(X2)
}

}
\concept{linalg}
