% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{cumsum.caracas_symbol}
\alias{cumsum.caracas_symbol}
\title{Cumulative Sums}
\usage{
\method{cumsum}{caracas_symbol}(x)
}
\arguments{
\item{x}{Elements to sum}
}
\description{
Cumulative Sums
}
\examples{
if (has_sympy()) {
  A <- matrix(1:9, 3)
  cumsum(A)
  B <- matrix_sym(3, 3)
  cumsum(B)
  C <- vector_sym(3)
  cumsum(C)
}

}
\concept{vectors}
