% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{dd}
\alias{dd}
\title{Symbolic differentiation of an expression}
\usage{
dd(expr, vars)
}
\arguments{
\item{expr}{A \code{caracas_symbol}}

\item{vars}{variables to take derivate with respect to}
}
\description{
Symbolic differentiation of an expression
}
\concept{calculus}
