% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{dd2}
\alias{dd2}
\title{Symbolic differentiation of second order of an expression}
\usage{
dd2(expr, vars)
}
\arguments{
\item{expr}{A \code{caracas_symbol}}

\item{vars}{variables to take derivate with respect to}
}
\description{
Symbolic differentiation of second order of an expression
}
\concept{caracas_symbol}
