% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{def_sym_vec}
\alias{def_sym_vec}
\title{Define symbol for components in vector}
\usage{
def_sym_vec(x, env = parent.frame())
}
\arguments{
\item{x}{The environment in which the assignment is made.}
}
\description{
Define symbol for components in vector
}
\examples{
if (has_sympy()) {
  def_sym(z1, z2, z3)
  u <- paste0("u", seq_len(3))
  ## Creates symbols u1, u2, u3 and binds to names u1, u2, u3 in R.
  def_sym_vec(u) 
  ## Same as (but easier than)
  def_sym(u1, u2, u3)
  ## Notice: this creates matrix [u1, u2, u3]
  as_sym(u)
 }
}
