% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{der2}
\alias{der2}
\title{Symbolic differentiation of second order of an expression}
\usage{
der2(expr, vars)
}
\arguments{
\item{expr}{A \code{caracas_symbol}}

\item{vars}{variables to take derivate with respect to}
}
\description{
Symbolic differentiation of second order of an expression
}
\examples{
x <- symbol("x")
y <- symbol("y")
f <- 3*x^2 + x*y^2
der2(f, x)

}
\concept{calculus}
