% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg_advanced.R
\name{determinant-caracas}
\alias{determinant-caracas}
\alias{determinant,caracas_symbol,logical-method}
\alias{determinant,caracas_symbol,missing-method}
\title{Determinant methods for caracas_symbol}
\usage{
\S4method{determinant}{caracas_symbol,logical}(x, logarithm = TRUE, ...)

\S4method{determinant}{caracas_symbol,missing}(x, logarithm = TRUE, ...)
}
\arguments{
\item{x}{A \code{caracas_symbol}.}

\item{logarithm}{Logical; if \code{TRUE}, return a log-determinant representation.}

\item{...}{Unused.}
}
\value{
An object of class \code{"det"} with components \code{modulus}, \code{sign},
and \code{logarithm}.
}
\description{
Symbolic determinant (and log-determinant) for \code{caracas_symbol}.
}
\concept{caracas_symbol}
