% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-math.R
\name{determinant.caracas_symbol}
\alias{determinant.caracas_symbol}
\title{Calculate the Determinant of a Matrix}
\usage{
\method{determinant}{caracas_symbol}(x, ...)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{\dots}{Not used}
}
\description{
Note that there is no argument for \code{logarithm} as with the generic
method.
}
\concept{linalg}
