% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{diag<-.caracas_symbol}
\alias{diag<-.caracas_symbol}
\title{Replace diagonal}
\usage{
\method{diag}{caracas_symbol}(x) <- value
}
\arguments{
\item{x}{A \code{caracas_symbol}.}

\item{value}{Replacement value}
}
\description{
Replace diagonal
}
\examples{
if (has_sympy()) {
  A <- matrix(c("a", 0, 0, 0, "a", "a", "a", 0, 0), 3, 3)
  B <- as_sym(A)
  B
  diag(B)
  diag(B) <- "b"
  B
  diag(B)
}

}
\concept{vectors}
