% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{diag-set}
\alias{diag-set}
\alias{diag<-}
\title{Replace matrix diagonal}
\usage{
diag(x) <- value
}
\arguments{
\item{x}{Object \code{x}}

\item{value}{Replacement value}
}
\description{
Replace matrix diagonal
}
\concept{linalg}
