% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{diag.caracas_symbol}
\alias{diag.caracas_symbol}
\title{Matrix diagonal}
\usage{
\method{diag}{caracas_symbol}(x, ...)
}
\arguments{
\item{x}{Object \code{x}}

\item{\dots}{Not used}
}
\description{
Matrix diagonal
}
\concept{linalg}
