% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{diag_sym}
\alias{diag_sym}
\title{Symbolic diagonal matrix}
\usage{
diag_sym(x, n = 1L, declare_symbols = TRUE, ...)
}
\arguments{
\item{x}{Character vector with diagonal}

\item{n}{Number of times \code{x} should be repeated}

\item{declare_symbols}{Passed on to \code{as_sym()} when constructing symbolic matrix}

\item{\dots}{Passed on to \code{rep(x, n, ...)}}
}
\description{
Symbolic diagonal matrix
}
\examples{
if (has_sympy()) {
  diag_sym(c("a", "b", "c"))
  diag_sym("a", 2)
}

}
\concept{linalg}
