% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{diff_mat}
\alias{diff_mat}
\title{Difference matrix}
\usage{
diff_mat(N, l = "-1", d = 1)
}
\arguments{
\item{N}{Number of rows (and columns)}

\item{l}{Value / symbol below main diagonal}

\item{d}{Value / symbol on main diagonal}
}
\description{
Difference matrix
}
\examples{
if (has_sympy()){
Dm <- diff_mat(4)
Dm
y <- vector_sym(4, "y")
Dm \%*\% y
}
}
