% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{dim<-.caracas_symbol}
\alias{dim<-.caracas_symbol}
\title{Dimensions of a caracas symbol}
\usage{
\method{dim}{caracas_symbol}(x) <- value
}
\arguments{
\item{x}{caracas symbol}

\item{value}{new dimension}
}
\description{
Dimensions of a caracas symbol
}
\examples{
if (has_sympy()) {
  v <- vector_sym(4)
  v
  dim(v)
  dim(v) <- c(2, 2)
  v
  m <- matrix_sym(2, 2)
  dim(m)
  dim(m) <- c(4, 1)
  m
}

}
\concept{linalg}
