% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{dim.caracas_symbol}
\alias{dim.caracas_symbol}
\title{Dimensions of a caracas symbol}
\usage{
\method{dim}{caracas_symbol}(x)
}
\arguments{
\item{x}{caracas symbol}
}
\description{
Dimensions of a caracas symbol
}
\concept{linalg}
