% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{doit}
\alias{doit}
\title{Perform calculations setup previously}
\usage{
doit(x)
}
\arguments{
\item{x}{A \code{caracas_symbol}}
}
\description{
Perform calculations setup previously
}
\concept{caracas_symbol}
