% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{eigenvals}
\alias{eigenvals}
\title{Eigenvalues}
\usage{
eigenvals(x)
}
\arguments{
\item{x}{Matrix to find eigenvalues for}
}
\description{
Eigenvalues
}
\examples{
A <- matrix(c("a", 0, 0, 0, "a", "a", "a", 0, 0), 3, 3)
B <- as_symbol(A)
eigenvals(B)
eigenvects(B)
eigen(eval(as_r(B), list(a = 2)))

}
\concept{linalg}
