% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{eigenvec}
\alias{eigenvec}
\title{Eigenvectors and eigenvalues}
\usage{
eigenvec(x)
}
\arguments{
\item{x}{Matrix to find eigenvectors and eigenvalues for}
}
\value{
Returns a list of eigenvectors where each entry
contains the eigenvector, eigenvalue and multiplicity
}
\description{
Eigenvectors and eigenvalues
}
\examples{
if (has_sympy()) {
  A <- matrix(c("a", "b", "c", "d"), 2, 2) \%>\% as_sym()
  evec <- eigenvec(A)
  evec
  evec1 <- evec[[1]]$eigvec
  evec1
  simplify(evec1)
  
  lapply(evec, function(l) simplify(l$eigvec))
  
  eigenval(A)
}

}
\concept{linalg}
