% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{eval_to_symbol}
\alias{eval_to_symbol}
\title{Create a symbol from a string}
\usage{
eval_to_symbol(x)
}
\arguments{
\item{x}{String to evaluate}
}
\value{
A \code{caracas_symbol}
}
\description{
Create a symbol from a string
}
\examples{
if (has_sympy()) {
   x <- symbol('x')
   (1+1)*x^2
   lim(sin(x)/x, "x", 0)
}

}
\concept{lowlevel}
