% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{expand}
\alias{expand}
\title{Expand expression}
\usage{
expand(x, ...)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{\dots}{Pass on to SymPy's expand, e.g. \code{force = TRUE}}
}
\description{
Expand expression
}
\examples{
if (has_sympy()) {
  def_sym(x)
  y <- log(exp(x))
  simplify(y)
  expand(simplify(y))
  expand(simplify(y), force = TRUE)
  expand_log(simplify(y))
}

}
\concept{simplify}
