% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{expand_func}
\alias{expand_func}
\title{Expand a function expression}
\usage{
expand_func(x)
}
\arguments{
\item{x}{A \code{caracas_symbol}}
}
\description{
Expand a function expression
}
\concept{simplify}
