% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{expand_log}
\alias{expand_log}
\title{Expand a logarithmic expression}
\usage{
expand_log(x)
}
\arguments{
\item{x}{A \code{caracas_symbol}}
}
\description{
Note that \code{force} as described at
\url{https://docs.sympy.org/latest/tutorial/simplification.html#expand-log} is used
meaning that some assumptions are taken.
}
\examples{
if (have_sympy()) {
  x <- symbol('x')
  y <- symbol('y')
  z <- log(x*y)
  z
  expand_log(z)
}

}
\concept{simplify}
