% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{expand_trig}
\alias{expand_trig}
\title{Expand a trigonometric expression}
\usage{
expand_trig(x)
}
\arguments{
\item{x}{A \code{caracas_symbol}}
}
\description{
Expand a trigonometric expression
}
\concept{caracas_symbol}
