% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{factor_}
\alias{factor_}
\title{Expand expression}
\usage{
factor_(x)
}
\arguments{
\item{x}{A \code{caracas_symbol}}
}
\description{
Expand expression
}
\examples{

if (has_sympy()){
  def_sym(x, y, z)
  factor_(x**3 - x**2 + x - 1)
  factor_(x**2*z + 4*x*y*z + 4*y**2*z)
}

}
\concept{simplify}
