% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{free_symbols}
\alias{free_symbols}
\title{Get free symbol in expression}
\usage{
free_symbols(x)
}
\arguments{
\item{x}{Expression in which to get the free symbols in}
}
\description{
Get free symbol in expression
}
\examples{
if (has_sympy()) {
  def_sym(a, b)
  x <- (a-b)^4
  free_symbols(x)
}
 
}
\concept{caracas_symbol}
