% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{get_basis}
\alias{get_basis}
\title{Get basis}
\usage{
get_basis(x)
}
\arguments{
\item{x}{caracas vector / matrix}
}
\description{
Get basis
}
\examples{
if (has_sympy()) {
  x <- vector_sym(3)
  get_basis(x)

  W <- matrix(c("r_1", "r_1", "r_2", "r_2", "0", "0", "u_1", "u_2"), nrow=4)
  W <- as_sym(W)
  get_basis(W)
}
}
\concept{linalg}
