% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{get_py}
\alias{get_py}
\title{Access 'py' object}
\usage{
get_py()
}
\value{
The 'py' object with direct access to the library.
}
\description{
Get the 'py' object.
Note that it gives you extra responsibilities
when you choose to access the 'py' object directly.
}
\examples{
if (has_sympy()) {
  py <- get_py()
}

}
\concept{sympy}
