% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{get_sympy}
\alias{get_sympy}
\title{Access 'SymPy' directly}
\usage{
get_sympy()
}
\value{
The 'SymPy' object with direct access to the library.
}
\description{
Get the 'SymPy' object.
Note that it gives you extra responsibilities
when you choose to access the 'SymPy' object directly.
}
\examples{
if (has_sympy()) {
  sympy <- get_sympy()
  sympy$solve("x**2-1", "x")
}

}
\concept{sympy}
