% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-math.R
\name{\%*\%.caracas_symbol}
\alias{\%*\%.caracas_symbol}
\title{Matrix multiplication}
\usage{
\method{\%*\%}{caracas_symbol}(x, y)
}
\arguments{
\item{x}{Object \code{x}}

\item{y}{Object \code{y}}
}
\description{
Matrix multiplication
}
\concept{linalg}
