% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{has_sympy}
\alias{has_sympy}
\title{Check if 'SymPy' is available}
\usage{
has_sympy()
}
\value{
\code{TRUE} if 'SymPy' is available, else \code{FALSE}
}
\description{
Check if 'SymPy' is available
}
\examples{
has_sympy()

}
\concept{sympy}
