% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_caracas_sorenh.R
\name{head_tail_caracas}
\alias{head_tail_caracas}
\alias{head.caracas_symbol}
\alias{tail.caracas_symbol}
\title{Head and tail for caracas matrices}
\usage{
\method{head}{caracas_symbol}(x, n = 6, ...)

\method{tail}{caracas_symbol}(x, n = 6, ...)
}
\arguments{
\item{x}{A caracas object}

\item{n}{Number of rows to be extracted}

\item{...}{Not used}
}
\value{
An object of same class as x
}
\description{
Head and tail for caracas matrices
}
\author{
Søren Højsgaard
}
