% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{install_sympy}
\alias{install_sympy}
\title{Install 'SymPy'}
\usage{
install_sympy(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method.
By default, "auto" automatically finds a method that will work
in the local environment.
Change the default to force a specific installation method.
Note that the "virtualenv" method is not available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda
using the PATH and other conventional install locations).}
}
\value{
None
}
\description{
Install the 'SymPy' Python package into a
virtual environment or Conda environment.
}
\concept{sympy}
