% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{inv_2x2}
\alias{inv_2x2}
\title{Inverse of 2x2 matrix}
\usage{
inv_2x2(x)
}
\arguments{
\item{x}{}
}
\value{
Columns of result.
The reason it is not a matrix is because SymPy has problems as 'I' both means ide
}
\description{
Finds the inverse of\preformatted{[A  B]
[C  D]
}
}
\details{
Arguments by row.
}
\examples{
if (has_sympy()) {
   W <- matrix_symbol("W") # 1 x 1 by default
   J <- sympy_declare("J", "Identity(1)") # Instead of I to avoid confusion with imaginary unit
   O <- sympy_declare("O", "ZeroMatrix(1, 1)")
   L <- matrix(c("J", "-W", "0", "J"), nrow = 2) \%>\%
     as_sym(declare_symbols = FALSE)
   L
   #inv(L): SymPy cannot figure it out...
   Linv <- inv_2x2(L)
   Linv
   
   t(Linv) \%*\% Linv
   O2 <- sympy_declare("O", "ZeroMatrix(2, 2)")
   A <- O2 \%*\% t(Linv) \%*\% Linv
   A
   dim(A)
}

}
\concept{caracas_symbol}
