% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg_advanced.R
\name{inv_block}
\alias{inv_block}
\title{Inverse of block matrix}
\usage{
inv_block(x)
}
\arguments{
\item{x}{A block matrix, either a caracas matrix, or a dense or a sparse matrix.}
}
\value{
The inverse in the same form as x.
}
\description{
Inverse of block matrix
}
\examples{
if (has_sympy()) {
 I <- diag_(1, 3)
 M <- as_sym(matrix(c("a", "b", "c", "d"), nrow=2))
 IM <- kronecker(I, M)
 inv(IM)
 inv_block(IM)

 IM. <- subs(IM, c("a", "b", "c", "d"), c(2,1,2,2))
 inv_block(IM.)
}
}
\author{
Søren Højsgaard
}
