% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg_advanced.R
\name{inv_woodbury}
\alias{inv_woodbury}
\title{Inverse using woodburys matrix identity}
\usage{
inv_woodbury(A, U, C, V = t(U), method = "ge", simplify = TRUE)
}
\arguments{
\item{A, U, C, V}{Either a caracas matrix, or a dense or a sparse matrix.}

\item{method}{One of the methods that can be supplied to inv().}

\item{simplify}{Should temporary quantities be simplified?}
}
\value{
The inverse of (A + U C V)
}
\description{
Computes the inverse of (A + U C V) provided that the inverse of A and C exists.
}
\examples{
if (has_sympy()) {

n <- 8
m <- 4
A <- diag_("a", n)
C <- diag_("c", m)
set.seed(1234)
U <- round(10 * (matrix(rnorm(n * m), nrow=n)))
U[U < 0] <- 0
U <- as_sym(U)
V <- t(U)

B <- A + U \%*\% C \%*\% V
B

system.time(Bi <- inv_woodbury(A, U, C, simplify=FALSE))
}
}
\author{
Søren Højsgaard
}
