% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{is_sym_list}
\alias{is_sym_list}
\title{Is list of caracas symbols}
\usage{
is_sym_list(x)
}
\arguments{
\item{x}{Object}
}
\value{
logical
}
\description{
Is list of caracas symbols
}
\author{
Søren Højsgaard
}
