% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{jacobian}
\alias{jacobian}
\title{Compute Jacobian}
\usage{
jacobian(expr, vars)
}
\arguments{
\item{expr}{'caracas expression'.}

\item{vars}{variables to take derivative with respect to}
}
\description{
Compute Jacobian
}
\examples{
if (has_sympy()) {
 x <- paste0("x", seq_len(3))
 def_sym_vec(x)
 y1 <- x1 + x2
 y2 <- x1^2 + x3
 y <- c(y1, y2)
 jacobian(y, x)
 u <- 2 + 4*x1^2
 jacobian(u, x1)
}
}
\seealso{
\code{\link[=score]{score()}}, \code{\link[=hessian]{hessian()}} \code{\link[=der]{der()}}
}
