% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_caracas_sorenh.R
\name{length_}
\alias{length_}
\title{Length of (caracas) object}
\usage{
length_(x)
}
\arguments{
\item{x}{an object}
}
\value{
length
}
\description{
Length of (caracas) object
}
\author{
Søren Højsgaard
}
