% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{lim}
\alias{lim}
\title{Limits}
\usage{
lim(f, var, val, dir, doit = TRUE, ...)
}
\arguments{
\item{f}{Function to take limit of}

\item{var}{Variable to take limit for}

\item{val}{Value for \code{var} to approach}

\item{dir}{Direction from where \code{var} should approach \code{val}: \code{'+'} or \code{'-'}}

\item{doit}{Evaluate the limit immediately (or later with \code{\link[=doit]{doit()}})}

\item{\dots}{Not used.}
}
\description{
Limits
}
\examples{
if (have_sympy()) {
  x <- symbol("x")
  lim(sin(x)/x, "x", 0)
  lim(1/x, "x", 0, dir = '+')
  lim(1/x, "x", 0, dir = '-')
}

}
\concept{caracas_symbol}
