% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{listify}
\alias{listify}
\title{Convert object to list of elements}
\usage{
listify(x)
}
\arguments{
\item{x}{Object}
}
\description{
Convert object to list of elements
}
\examples{
if (has_sympy()) {
  x <- as_sym("Matrix([[b1*x1/(b2 + x1)], [b1*x2/(b2 + x2)], [b1*x3/(b2 + x3)]])")
  listify(x)
  
  xT <- t(x)
  listify(xT)
}

}
\concept{caracas_symbol}
