% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{mat_div_mult}
\alias{mat_div_mult}
\alias{mat_factor_div}
\alias{mat_factor_mult}
\title{Divide or multiply matrix with factor.}
\usage{
mat_factor_div(m, s)

mat_factor_mult(m, s)
}
\arguments{
\item{m}{Matrix}

\item{s}{Factor}
}
\description{
Divide or multiply matrix with factor.
}
\concept{caracas_symbol}
