% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{mat_pow}
\alias{mat_pow}
\title{Matrix power}
\usage{
mat_pow(x, pow = "1")
}
\arguments{
\item{x}{A \code{caracas_symbol}, a matrix.}

\item{pow}{Power to raise matrix \code{x} to}
}
\description{
Matrix power
}
\examples{
if (has_sympy()) {
  M <- matrix_(c("1", "a", "a", 1), 2, 2)
  M
  mat_pow(M, 1/2)
}

}
\concept{linalg}
