% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops_math.R
\name{matrix-products}
\alias{matrix-products}
\alias{\%*\%}
\alias{\%*\%.caracas_symbol}
\title{Matrix multiplication}
\usage{
x \%*\% y

\method{\%*\%}{caracas_symbol}(x, y)
}
\arguments{
\item{x}{Object \code{x}}

\item{y}{Object \code{y}}
}
\description{
Matrix multiplication

Matrix multiplication
}
\seealso{
\code{\link[base:matmult]{base::\%*\%()}}

\code{\link[base:matmult]{base::\%*\%()}}
}
\concept{linalg}
