% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{matrix_}
\alias{matrix_}
\title{Symbolic matrix}
\usage{
matrix_(..., declare_symbols = TRUE)
}
\arguments{
\item{\dots}{Passed on to \code{\link[=matrix]{matrix()}}}

\item{declare_symbols}{Passed on to \code{as_sym()} when constructing symbolic matrix}
}
\description{
Symbolic matrix
}
\examples{
if (has_sympy()) {
  matrix_(1:9, nrow = 3)
  matrix_("a", 2, 2)
}

}
\concept{linalg}
