% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{matrix_cross_product}
\alias{matrix_cross_product}
\alias{crossprod_}
\alias{tcrossprod_}
\alias{crossprod,caracas_symbol,missing-method}
\alias{tcrossprod,caracas_symbol,missing-method}
\title{Matrix cross product}
\usage{
crossprod_(x, y = NULL)

tcrossprod_(x, y = NULL)

\S4method{crossprod}{caracas_symbol,missing}(x, y)

\S4method{tcrossprod}{caracas_symbol,missing}(x, y)
}
\arguments{
\item{x, y}{caracas matrices}
}
\description{
Matrix cross product
}
\examples{
if (has_sympy()) {
  s  <- as_sym("[[r1, r2, r3], [u1, u2, u3]]")
  s2 <- apply(as_character_matrix(s), 2, function(x) (paste("1/(", x, ")")))
  as_sym(s2)
}

}
\concept{linalg}
