% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{matrix_symbol}
\alias{matrix_symbol}
\title{Create matrix symbol}
\usage{
matrix_symbol(x, nrow = 1L, ncol = 1L)
}
\arguments{
\item{x}{Name of matrix symbol}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}
}
\description{
Create matrix symbol
}
\examples{
if (has_sympy()) {
   A <- matrix_symbol("A", 3, 3)
   A
   dim(A)
   t(A)
   inv(A)
}

}
\seealso{
\code{\link[=sympy_declare]{sympy_declare()}}
}
\concept{caracas_symbol}
