% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_eval.R
\name{N}
\alias{N}
\title{Numerical evaluation}
\usage{
N(x, digits = 15)
}
\arguments{
\item{x}{caracas object}

\item{digits}{Number of digits}
}
\description{
Numerical evaluation
}
\examples{
if (has_sympy()) {
  n_2 <- as_sym("2")
  n_pi <- as_sym("pi", declare_symbols = FALSE)
  x <- sqrt(n_2) * n_pi
  x
  N(x)
  N(x, 5)
  N(x, 50)
  as.character(N(x, 50))
}

}
