% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-math.R
\name{Ops.caracas_symbol}
\alias{Ops.caracas_symbol}
\title{Math operators}
\usage{
\method{Ops}{caracas_symbol}(e1, e2)
}
\arguments{
\item{e1}{A \code{caracas_symbol}.}

\item{e2}{A \code{caracas_symbol}.}
}
\description{
Math operators
}
