% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_sym.R
\name{optim_sym}
\alias{optim_sym}
\title{General purpose optimization of caracas symbol.}
\usage{
optim_sym(
  par,
  fn,
  gr = NULL,
  ...,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  lower = -Inf,
  upper = Inf,
  control = list(),
  hessian = FALSE
)
}
\arguments{
\item{par}{As for optim()}

\item{fn}{A caracas symbol}

\item{gr}{As for optim()}

\item{...}{As for optim()}

\item{method}{As for optim()}

\item{lower}{As for optim()}

\item{upper}{As for optim()}

\item{control}{As for optim()}

\item{hessian}{As for optim()}
}
\value{
As for optim()
}
\description{
Adaptation of optim() to work on caracas symbols.
}
\details{
The caracas symbol \code{fn} is coerced into an R function on which optim() is subsequently applied.
}
\author{
Søren Højsgaard
}
