% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\alias{as_piecewise}
\alias{piecewise_cond}
\alias{piecewise_expr}
\alias{[.caracas_piecewise}
\title{Create a piecewise object}
\usage{
as_piecewise(pw)

piecewise_cond(pw)

piecewise_expr(pw)

\method{[}{caracas_piecewise}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{pw}{A caracas object containg a Piecewise specification}

\item{x}{A \code{caracas_symbol}.}

\item{i}{row indices specifying elements to extract or replace}

\item{j}{column indices specifying elements to extract or replace}

\item{\dots}{Not used}

\item{drop}{Simplify dimensions of resulting object}
}
\value{
A list
}
\description{
Create a piecewise object

Extract or replace parts of an object
}
\examples{
if (has_sympy()) {
  library(caracas)
  def_sym(r, n, j)
  sum1 <- sum_(r^j, var=j, lower=0, upper=n)
  pw <- as_piecewise(sum1)
  pw[[2]]$expr

  pw |> piecewise_cond()
  pw |> piecewise_expr()

}

}
\author{
Søren Højsgaard
}
