% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_scaled.R
\name{print.caracas_scaled_matrix}
\alias{print.caracas_scaled_matrix}
\title{Print scaled matrix}
\usage{
\method{print}{caracas_scaled_matrix}(x, ...)
}
\arguments{
\item{x}{A \code{caracas_scaled_matrix}}

\item{\dots}{Passed to \code{\link[=print.caracas_symbol]{print.caracas_symbol()}}}
}
\description{
Print scaled matrix
}
\concept{output}
