% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{prod_}
\alias{prod_}
\title{Product of a function}
\usage{
prod_(f, var, lower, upper, doit = TRUE)
}
\arguments{
\item{f}{Function to take product of}

\item{var}{Variable to take product for (either string or \code{caracas_symbol})}

\item{lower}{Lower limit}

\item{upper}{Upper limit}

\item{doit}{Evaluate the product immediately (or later with \code{\link[=doit]{doit()}})}
}
\description{
Product of a function
}
\examples{
if (has_sympy()) {
  x <- symbol("x")
  p <- prod_(1/x, "x", 1, 10)
  p
  as_expr(p)
  prod(1/(1:10))
  n <- symbol("n")
  prod_(x, x, 1, n)
}

}
\concept{calculus}
