% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{QRdecomposition}
\alias{QRdecomposition}
\title{QR decomposition}
\usage{
QRdecomposition(x)
}
\arguments{
\item{x}{Matrix to find QR decomposition for}
}
\value{
Returns a list of eigenvectors where each entry
contains the eigenvector, eigenvalue and multiplicity
}
\description{
QR decomposition
}
\examples{
if (has_sympy()) {
  A <- matrix(c("a", "0", "0", "1"), 2, 2) \%>\% as_sym()
  qr_res <- QRdecomposition(A)
  qr_res$Q
  qr_res$R
}

}
\concept{linalg}
