% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_as_sym.R
\name{re_sym}
\alias{re_sym}
\title{Refresh as symbol}
\usage{
re_sym(x)
}
\arguments{
\item{x}{caracas symbol}
}
\value{
caracas symbol
}
\description{
Refresh as symbol
}
\examples{
if (has_sympy()) {
x <- symbol('x', positive=T)
e <- 2*x^2
## Substitute 2 for x
subs(e, "x", 2) 
## Fails because of restriction that x is positive

## Refresh symbol 
e <- re_sym(e)
## Works
subs(e, "x", "2") ## OK
}
}
