% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{rootsolve}
\alias{rootsolve}
\title{Find the roots in a system of non-linear equations}
\usage{
rootsolve(lhs, vars)
}
\arguments{
\item{lhs}{Equation (or equations as row vector/1xn matrix)}

\item{vars}{vector of variable names or symbols}
}
\description{
Find the roots in a system of non-linear equations
}
\concept{solve}
