% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{rowSums_colSums}
\alias{rowSums_colSums}
\alias{rowSums_}
\alias{colSums_}
\title{Form Row and Column Sums}
\usage{
rowSums_(x)

colSums_(x)
}
\arguments{
\item{x}{Symbolic matrix}
}
\description{
Form Row and Column Sums
}
\examples{
if (has_sympy()) {
  X <- matrix_(paste0("x_",c(1,1,1,1,2,2,2,2,3,4,3,4)), nrow=4)
  rowSums_(X)
  colSums_(X)
}

}
\concept{linalg}
