% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{set_print_type}
\alias{set_print_type}
\title{Set how printing is done}
\usage{
set_print_type(type = c("pretty_utf8", "string"))
}
\arguments{
\item{type}{'pretty_utf8', 'string'}
}
\description{
Set how printing is done
}
