% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_triangular}
\alias{solve_triangular}
\alias{solve_lower_triangular}
\alias{solve_upper_triangular}
\title{Solve lower or upper triangular system}
\usage{
solve_lower_triangular(a, b, ...)

solve_upper_triangular(a, b, ...)
}
\arguments{
\item{a}{caracas_symbol}

\item{b}{If provided, either a caracas_symbol (if not, \code{as_sym()} is called on the object)}

\item{\dots}{Not used}
}
\description{
Solve lower or upper triangular system
}
\examples{
if (has_sympy()) {
  A <- matrix_sym(3, 3)
  A[upper.tri(A)] <- 0
  solve_lower_triangular(A) |> simplify()
  A <- matrix_sym(3, 3)
  A[lower.tri(A)] <- 0
  solve_upper_triangular(A) |> simplify()
}

}
\concept{solve}
