% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{special_matrices}
\alias{special_matrices}
\alias{zeros_sym}
\alias{ones_sym}
\alias{eye_sym}
\title{Special matrices: zeros_sym, ones_sym, eye_sym}
\usage{
zeros_sym(nrow, ncol)

ones_sym(nrow, ncol)

eye_sym(nrow, ncol)
}
\arguments{
\item{nrow, ncol}{Number of rows and columns of output}
}
\description{
Special matrices: zeros_sym, ones_sym, eye_sym
}
\examples{

if (has_sympy()){
  zeros_sym(3, 4)
  ones_sym(3, 4)
  eye_sym(3, 4)
}

}
\seealso{
\code{\link[=diag_]{diag_()}}, \code{\link[=matrix_sym]{matrix_sym()}}, \code{\link[=vector_sym]{vector_sym()}}
}
\concept{linalg}
