% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{subs_list}
\alias{subs_list}
\title{Substitute symbol for value}
\usage{
subs_list(sym_list, nms_list, vls_list)
}
\arguments{
\item{sym_list}{Expression}

\item{nms_list}{Names of symbols (see Details)}

\item{vls_list}{Values that \code{nms} is substituted with (see Details)}
}
\description{
Substitute symbol for value
}
