% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{subs_mat}
\alias{subs_mat}
\title{Substitute af vector of symbols for a vector of values in matrix.}
\usage{
subs_mat(s, x, v)
}
\arguments{
\item{s}{Matrix}

\item{x}{Names of symbol (vector)}

\item{v}{Values for \code{x} (vector)

#' @examples
if (has_sympy()) {
S <- matrix_sym(2,2)
x <- c("v11", "v12")
v <- c(2,4)
subs_mat(S, x, v)
}}
}
\description{
Substitute af vector of symbols for a vector of values in matrix.
}
