% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{subs_single}
\alias{subs_single}
\title{Substitute symbol for value}
\usage{
subs_single(s, x, v)
}
\arguments{
\item{s}{Expression}

\item{x}{Name of symbol (character)}

\item{v}{Value for \code{x}}
}
\description{
Substitute symbol for value
}
\examples{
if (has_sympy()) {
   x <- symbol('x')
   e <- 2*x^2
   e
   subs_single(e, "x", "2")
   y <- as_sym("2")
   subs_single(e, "x", y)
   b1 <- symbol("a"); str(b1)
   b2 <- b1 + 1; str(b2)
   b3 <- subs(b2, "a", "k");
   str(b3)
}

}
\seealso{
\link{subs}, \code{\link[=subs_lst]{subs_lst()}}
}
\concept{caracas_symbol}
