% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{subs_vec}
\alias{subs_vec}
\title{Substitute af vector of symbols for a vector of values}
\usage{
subs_vec(s, x, v)
}
\arguments{
\item{s}{Expression}

\item{x}{Names of symbol (vector)}

\item{v}{Values for \code{x} (vector)}
}
\description{
Substitute af vector of symbols for a vector of values
}
\examples{
if (has_sympy()) {
   x <- as_sym(paste0('x', 1:3))
   e <- 2*x^2
   e
   subs_vec(e, x, 1:3)
   subs_vec(e, x, x^2)
}

}
\seealso{
\code{\link[=subs]{subs()}}, \code{\link[=subs_lst]{subs_lst()}}
}
\concept{caracas_symbol}
