% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[<-.caracas_symbol}
\alias{[<-.caracas_symbol}
\title{Extract or replace parts of an object}
\usage{
\method{[}{caracas_symbol}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A \code{caracas_symbol}.}

\item{i}{row indices specifying elements to extract or replace}

\item{j}{column indices specifying elements to extract or replace}

\item{\dots}{Not used}

\item{value}{Replacement value}
}
\description{
Extract or replace parts of an object
}
\examples{
if (has_sympy()) {
  A <- matrix(c("a", 0, 0, 0, "a", "a", "a", 0, 0), 3, 3)
  B <- as_sym(A)
  B[, 2] <- "x"
  B
}

}
\concept{vectors}
