% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{sum.caracas_symbol}
\alias{sum.caracas_symbol}
\title{Summation}
\usage{
\method{sum}{caracas_symbol}(..., na.rm = FALSE)
}
\arguments{
\item{\dots}{Elements to sum}

\item{na.rm}{Not used}
}
\description{
Summation
}
\concept{vectors}
