% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{sumf}
\alias{sumf}
\title{Sum of a function}
\usage{
sumf(f, var, lower, upper, doit = TRUE)
}
\arguments{
\item{f}{Function to take sum of}

\item{var}{Variable to take sum for (either string or \code{caracas_symbol})}

\item{lower}{Lower limit}

\item{upper}{Upper limit}

\item{doit}{Evaluate the sum immediately (or later with \code{\link[=doit]{doit()}})}
}
\description{
Sum of a function
}
\examples{
if (have_sympy()) {
  x <- symbol("x")
  s <- sumf(1/x, "x", 1, 10)
  as_expr(s)
  sum(1/(1:10))
  n <- symbol("n")
  simplify(sumf(x, x, 1, n))
}

}
\concept{calculus}
