% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{sym_inherits}
\alias{sym_inherits}
\title{Ask if type of caracas symbol is of a requested type}
\usage{
sym_inherits(x, what)
}
\arguments{
\item{x}{An object, a caracas object is expected}

\item{what}{Requested type (e.g. atomic, vector, list, matrix)}
}
\description{
Ask if type of caracas symbol is of a requested type
}
\concept{caracas_symbol}
