% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more-coerce.R
\name{symbol_class}
\alias{symbol_class}
\title{Ask type of caracas symbol}
\usage{
symbol_class(x)
}
\arguments{
\item{x}{An object, a caracas object is expected}
}
\description{
Ask type of caracas symbol
}
