% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{sympy_declare}
\alias{sympy_declare}
\title{Declare symbol manually by Python syntax}
\usage{
sympy_declare(x, cmd)
}
\arguments{
\item{x}{Name of symbol}

\item{cmd}{Command (SymPy/Python syntax)}
}
\description{
Declare symbol manually by Python syntax
}
\examples{
if (has_sympy()) {
   x <- sympy_declare("x", "Symbol('x')")
   x
   
   x <- sympy_declare("x", "Symbol('x', positive = True)")
   x
   ask(x, "positive")
   
   A <- matrix_symbol("A", 3, 3) # or: sympy_declare("A", "MatrixSymbol('A', 3, 3)")
   A
   dim(A)
   t(A)
   inv(A)
   
   A <- matrix_symbol("A", "MatrixSymbol('A', 3, 3)")
   O <- sympy_declare("O", "ZeroMatrix(3, 3)")
   A \%*\% O
}

}
\seealso{
\code{\link[=matrix_symbol]{matrix_symbol()}}
}
\concept{caracas_symbol}
