% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{sympy_version}
\alias{sympy_version}
\title{Get 'SymPy' version}
\usage{
sympy_version()
}
\value{
The version of the 'SymPy' available
}
\description{
Get 'SymPy' version
}
\examples{
if (has_sympy()) {
  sympy_version()
}

}
\concept{sympy}
