% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{systemsolve}
\alias{systemsolve}
\title{Solves a system of non-linear equations}
\usage{
systemsolve(lhs, rhs, vars)
}
\arguments{
\item{lhs}{Equation (or equations as row vector/1xn matrix)}

\item{rhs}{Equation (or equations as row vector/1xn matrix)}

\item{vars}{vector of variable names or symbols}
}
\description{
Finds roots in \code{lhs - rhs}.
}
\concept{solve}
