% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{t.caracas_symbol}
\alias{t.caracas_symbol}
\title{Transpose of matrix}
\usage{
\method{t}{caracas_symbol}(x)
}
\arguments{
\item{x}{If \code{caracas_symbol} treat as such, else
call \code{\link[base:t]{base::t()}}.}
}
\description{
Transpose of matrix
}
\concept{linalg}
