% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{taylor}
\alias{taylor}
\title{Taylor expansion}
\usage{
taylor(f, x0 = 0, n = 6)
}
\arguments{
\item{f}{Function to be expanded}

\item{x0}{Point to expand around}

\item{n}{Order of remainder term}
}
\description{
Taylor expansion
}
\examples{
if (has_sympy()) {
  def_sym(x)
  f <- cos(x)
  ft_with_O <- taylor(f, x0 = 0, n = 4+1)
  ft_with_O
  ft_with_O |> drop_remainder() |> as_expr()
}

}
\seealso{
\code{\link[=drop_remainder]{drop_remainder()}}
}
\concept{calculus}
