% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tex.caracas_factor_list}
\alias{tex.caracas_factor_list}
\title{Print factor list}
\usage{
\method{tex}{caracas_factor_list}(x, ...)
}
\arguments{
\item{x}{factor list}

\item{\dots}{Other arguments passed along}
}
\description{
Print factor list
}
\concept{output}
