% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_scaled.R
\name{tex.caracas_scaled_matrix}
\alias{tex.caracas_scaled_matrix}
\title{Export scaled matrix to tex}
\usage{
\method{tex}{caracas_scaled_matrix}(x, ...)
}
\arguments{
\item{x}{scaled matrix}

\item{\dots}{Other arguments passed along}
}
\description{
Export scaled matrix to tex
}
\concept{output}
