% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{tex.factor_list}
\alias{tex.factor_list}
\title{Print factor list}
\usage{
\method{tex}{factor_list}(x, ...)
}
\arguments{
\item{x}{factor list}

\item{\dots}{Other arguments passed along}
}
\description{
Print factor list
}
