% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tex_align}
\alias{tex_align}
\title{tex_align}
\usage{
tex_align(x, zero_as_dot = FALSE)
}
\arguments{
\item{x}{list of caracas symbol}

\item{zero_as_dot}{Print zero as dots}
}
\value{
latex string
}
\description{
tex_align
}
\author{
Søren Højsgaard
}
