% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tex_caracas}
\alias{tex_caracas}
\alias{tex_list}
\alias{tex_align}
\alias{tex_eq}
\title{Export caracas objects to TeX}
\usage{
tex_list(..., x = NULL, zero_as_dot = FALSE, matstr = NULL)

tex_align(x, zero_as_dot = FALSE, nonumber = TRUE)

tex_eq(x, zero_as_dot = FALSE)
}
\arguments{
\item{\dots}{Objects to be put in tex for. Can be \code{caracas_symbol}s
and other (simple) R objects (atomics, dataframes, matrices).}

\item{x}{caracas symbol}

\item{zero_as_dot}{Print zero as dots}

\item{matstr}{Replace \verb{\\begin\{matrix\}} with another environment,
e.g. \code{pmatrix}.  If vector of length two, the second element is
an optional argument.}

\item{nonumber}{If \code{TRUE}, add \verb{\\nonumber} to the end of each line}
}
\value{
latex string

latex string
}
\description{
Export caracas objects to TeX

tex_align

tex_eq
}
\examples{
if (has_sympy()) {
  X <- matrix_sym(4, 2, "a")
  b <- vector_sym(2, "b")
  y <- vector_sym(4, "y")

  tex(X)
  tex_list("X=", X)
  tex_list(y, "=", X, b)
  tex_list(x=list(y, "=", X, b))
  tex_eq(X)
  tex_align(list(list("X", X)))
  tex_align(list(list("X", X), list("b", b)))

  M <- iris[1:3, 1:2]
  tex_list(M, "+", M, "=", M + M)
  tex_list(x=list(M, "+", M, "=", M + M))

}
}
\author{
Søren Højsgaard
}
\concept{output}
