% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tex_eq}
\alias{tex_eq}
\title{tex_eq}
\usage{
tex_eq(x, zero_as_dot = FALSE)
}
\arguments{
\item{x}{caracas symbol}

\item{zero_as_dot}{Print zero as dots}
}
\value{
latex string
}
\description{
tex_eq
}
\author{
Søren Højsgaard
}
