% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{texdump}
\alias{texdump}
\title{Dump latex representation of sympy object.}
\usage{
texdump(x, name = "obj")
}
\arguments{
\item{x}{An object that can be put in latex format with caracas' tex() function.}

\item{name}{Name of tex file.}
}
\value{
Nothing, but a .tex file and a .pdf file is generated.
}
\description{
Dump latex representation of sympy object and compile document into pdf.
}
\examples{
S <- matrix_sym_symmetric(3, "s")
\dontrun{
texdump(S)
}

}
