% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{texshow}
\alias{texshow}
\title{Dump latex representation of sympy object.}
\usage{
texshow(x)
}
\arguments{
\item{x}{An object that can be put in latex format with caracas' tex()
function or a character string with tex code (in math mode).}
}
\value{
Nothing, but a .tex file and a .pdf file is generated.
}
\description{
Dump latex representation of sympy object and compile document into pdf.
}
\examples{
if (has_sympy()) {
S <- matrix_sym_symmetric(3, "s")
S
\dontrun{
texshow(S)
texshow(paste0("S = ", tex(S)))
}
}

}
\concept{output}
