% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more-coerce.R
\name{to_something}
\alias{to_something}
\alias{to_list}
\alias{to_vector}
\alias{to_matrix}
\title{Coerce caracas object}
\usage{
to_list(x)

to_vector(x)

to_matrix(x)
}
\arguments{
\item{x}{a caracas object is expected}
}
\description{
Coerce caracas object
}
