% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{toeplitz_}
\alias{toeplitz_}
\title{Topelitz matrix as caracas matrix.}
\usage{
toeplitz_(x, ...)
}
\arguments{
\item{x}{The first row of a toeplitz matrix}

\item{...}{Additional arguments passed on to toeplitz}
}
\description{
Topelitz matrix as caracas matrix.
}
\concept{linalg}
