% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{tuplify}
\alias{tuplify}
\title{Convert object to tuple}
\usage{
tuplify(x)
}
\arguments{
\item{x}{Object}
}
\description{
Convert object to tuple
}
\examples{
if (has_sympy()) {
  x <- as_sym("Matrix([[b1*x1/(b2 + x1)], [b1*x2/(b2 + x2)], [b1*x3/(b2 + x3)]])")
  tuplify(x)
  
}

}
\concept{caracas_symbol}
