% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{underscore}
\alias{underscore}
\alias{c_}
\alias{rep_}
\alias{seq_}
\title{caracas version of standard R functions}
\usage{
c_(...)

rep_(...)

seq_(...)
}
\arguments{
\item{...}{Arguments to be passed on}
}
\description{
a caracas object is returned
}
