% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{unique.caracas_symbol}
\alias{unique.caracas_symbol}
\title{Extract unique elements}
\usage{
\method{unique}{caracas_symbol}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A caracas vector or matrix}

\item{incomparables}{Same meaning as for other unique methods}

\item{...}{Additional arguments; currently not used.}
}
\description{
Extract unique elements
}
\examples{
if (has_sympy()){
v <- vector_sym(4)
v2 <- rep(v, each=2)
unique(v2)
}
}
