% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_scaled.R
\name{unscale_matrix}
\alias{unscale_matrix}
\title{Extract matrix from scaled matrix}
\usage{
unscale_matrix(X)
}
\arguments{
\item{X}{scaled matrix created with \code{\link[=scale_matrix]{scale_matrix()}}}
}
\description{
Extract matrix from scaled matrix
}
\examples{
if (has_sympy()) {
  V <- matrix_sym(2, 2, "v")
  a <- symbol("a")
  Ks <- scale_matrix(V, a, divide = FALSE)
  Ks
  unscale_matrix(Ks)
  V \%*\% a
}

}
\concept{linalg}
