% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_caracas_sorenh.R
\name{vec_along}
\alias{vec_along}
\title{Create a vector along}
\usage{
vec_along(x, entry = "v")
}
\arguments{
\item{x}{a caracas object}

\item{entry}{value}
}
\value{
a caracas vector
}
\description{
Create a vector along
}
\author{
Søren Højsgaard
}
