% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{as.character.caracas_symbol}
\alias{as.character.caracas_symbol}
\title{Convert symbol to character}
\usage{
\method{as.character}{caracas_symbol}(x, ...)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{\dots}{not used}
}
\description{
Convert symbol to character
}
\concept{output}
