% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{as_character_matrix}
\alias{as_character_matrix}
\title{Get matrix as character matrix}
\usage{
as_character_matrix(x)
}
\arguments{
\item{x}{caracas symbol}
}
\description{
Get matrix as character matrix
}
\examples{
if (has_sympy()) {
  s  <- as_sym("[[r1, r2, r3], [u1, u2, u3]]")
  s2 <- apply(as_character_matrix(s), 2, function(x) (paste("1/(", x, ")")))
  as_sym(s2)
}

}
\concept{linalg}
