% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{doit}
\alias{doit}
\title{Perform calculations setup previously}
\usage{
doit(x)
}
\arguments{
\item{x}{A \code{caracas_symbol}}
}
\description{
Perform calculations setup previously
}
\examples{
if (has_sympy()) {
   x <- symbol('x')
   res <- limf(sin(x)/x, "x", 0, doit = FALSE)
   res 
   doit(res)
}

}
\concept{caracas_symbol}
