% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{eigenval}
\alias{eigenval}
\title{Eigenvalues}
\usage{
eigenval(x)
}
\arguments{
\item{x}{Matrix to find eigenvalues for}
}
\description{
Eigenvalues
}
\examples{
if (have_sympy()) {
  A <- matrix(c("a", 0, 0, 0, "a", "a", "a", 0, 0), 3, 3)
  B <- as_symbol(A)
  eigenval(B)
  eigenvec(B)
  eigen(eval(as_r(B), list(a = 2)))
}

}
\concept{linalg}
