% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{eigenvec}
\alias{eigenvec}
\title{Eigenvectors and eigenvalues}
\usage{
eigenvec(x)
}
\arguments{
\item{x}{Matrix to find eigenvectors and eigenvalues for}
}
\description{
Eigenvectors and eigenvalues
}
\examples{
if (have_sympy()) {
  A <- matrix(c("a", 0, 0, 0, "a", "a", "a", 0, 0), 3, 3)
  B <- as_sym(A)
  eigenval(B)
  eigenvec(B)
  eigen(eval(as_expr(B), list(a = 2)))
}

}
\concept{linalg}
