% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_symbol.R
\name{fraction_parts}
\alias{fraction_parts}
\alias{numerator}
\alias{denominator}
\title{Get numerator and denominator of a fraction}
\usage{
fraction_parts(x)

numerator(x)

denominator(x)
}
\arguments{
\item{x}{Fraction}
}
\description{
Get numerator and denominator of a fraction
}
\examples{
if (has_sympy()) {
     x <- as_sym("a/b")
     frac <- fraction_parts(x)
     frac
     frac$numerator
     frac$denominator
 }

}
\concept{caracas_symbol}
