% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg-sh.R
\name{getLA}
\alias{getLA}
\title{getLA}
\usage{
getLA(x, slot, ...)
}
\arguments{
\item{x}{A matrix for which a property is requested}

\item{slot}{The property requested}

\item{...}{Auxillary arguments}
}
\value{
Returns the requested property of a matrix.
}
\description{
getLA
}
\examples{
if (has_sympy()) {
  A <- matrix(c("a", "0", "0", "1"), 2, 2) \%>\% as_sym()
  getLA(A, "QR")
  getLA(A, "eigenval")
  getLA(A, "eigenvec")
  getLA(A, "inv")
  getLA(A, "echelon_form")
  getLA(A, "rank")
  getLA(A, "det")
}

}
\concept{linalg}
