% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg_blockinv.R
\name{inv_woodbury}
\alias{inv_woodbury}
\title{Inverse using woodburys matrix identity}
\usage{
inv_woodbury(A, U, C, V = t(U), method = "ge", simplify = TRUE)
}
\arguments{
\item{A, U, C, V}{Either a caracas matrix, or a dense or a sparse matrix.}

\item{method}{One of the methods that can be supplied to inv().
\subsection{@param timing Should timing be printed}{
}}
}
\value{
The inverse of (A+UCV)
}
\description{
Computes the inverse of (A+UCV) provided that the inverse of A and C exists.
}
\examples{
if (has_sympy()) {

n <- 8
m <- 4
A <- diag_("a", n)
U <- round(10*(matrix(rnorm(n*m), nrow=n)))
U[U < 0] <- 0
U <- as_sym(U)
V <- t(U)
C <- diag_("c", m)

B <- A + U \%*\% C \%*\% V
B

Bi <- inv_woodbury(A, U, C)
}
}
\author{
Søren Højsgaard
}
