% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg_advanced.R
\name{linalg}
\alias{linalg}
\alias{columnspace}
\alias{nullspace}
\alias{rowspace}
\alias{orthcompspace}
\alias{intersectionspace}
\alias{leftnullspace}
\alias{singular_values}
\alias{inv}
\alias{eigenval}
\alias{eigenvec}
\alias{eigen_}
\alias{GramSchmidt}
\alias{pinv}
\alias{QRdecomposition}
\alias{qr,caracas_symbol-method}
\alias{qr.Q,QRdecomposition-method}
\alias{qr.R,QRdecomposition-method}
\alias{determinant,caracas_symbol,ANY-method}
\alias{LUdecomposition}
\alias{chol.caracas_symbol}
\alias{svd_}
\alias{det}
\alias{trace_}
\title{Do linear algebra operation}
\usage{
columnspace(x, matrix = TRUE)

nullspace(x, matrix = TRUE)

rowspace(x, matrix = TRUE)

orthcompspace(x, x2 = NULL)

intersectionspace(x, x2)

leftnullspace(x)

singular_values(x)

inv(x, method = c("ge", "gauss", "lu", "cf", "yac"))

eigenval(x)

eigenvec(x)

eigen_(x)

GramSchmidt(x)

pinv(x)

QRdecomposition(x)

\S4method{qr}{caracas_symbol}(x)

\S4method{qr.Q}{QRdecomposition}(qr)

\S4method{qr.R}{QRdecomposition}(qr)

\S4method{determinant}{caracas_symbol,ANY}(x, logarithm = TRUE, ...)

LUdecomposition(x)

\method{chol}{caracas_symbol}(x, ...)

svd_(x, ...)

det(x, ...)

trace_(x)
}
\arguments{
\item{x, x2}{A matrix for which a property is requested.}

\item{matrix}{When relevant should a matrix be returned.}

\item{method}{The default works by Gaussian elimination.  The
alternatives are $LU$ decomposition (\code{lu}), the cofactor
method (\code{cf}), and \code{Ryacas} (\code{yac}).}

\item{qr}{A QRdecomposition object.}

\item{logarithm}{logical.}

\item{...}{Auxillary arguments.}
}
\value{
Returns the requested property of a matrix.
}
\description{
Performs various linear algebra operations like finding the inverse,
the QR decomposition, the eigenvectors and the eigenvalues.
}
\examples{
if (has_sympy()) {
  A <- matrix(c("a", "0", "0", "1"), 2, 2) |> as_sym()
  
  QRdecomposition(A)
  LUdecomposition(A)
  #chol(A) # error
  chol(A, hermitian = FALSE)
  eigenval(A)
  eigenvec(A)
  inv(A)
  det(A)
  rowspace(A)
  columnspace(A)
  nullspace(A)
  
  ## Matrix inversion:
  d <- 3
  m <- matrix_sym(d, d)
  print(system.time(inv(m)))                  ## Gauss elimination
  print(system.time(inv(m, method="cf")))     ## Cofactor 
  print(system.time(inv(m, method="lu")))     ## LU decomposition
  if (requireNamespace("Ryacas")){
    print(system.time(inv(m, method="yac")))  ## Use Ryacas
  }

  A <- matrix(c("a", "b", "c", "d"), 2, 2) |> as_sym()
  evec <- eigenvec(A)
  evec
  evec1 <- evec[[1]]$eigvec
  evec1
  simplify(evec1)
  
  lapply(evec, function(l) simplify(l$eigvec))

  A <- as_sym("[[1, 2, 3], [4, 5, 6]]")
  pinv(A)
}

}
\seealso{
\code{\link[=do_la]{do_la()}}
}
\concept{linalg}
