% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops-math.R
\name{Math.caracas_symbol}
\alias{Math.caracas_symbol}
\title{Math functions}
\usage{
\method{Math}{caracas_symbol}(x, ...)
}
\arguments{
\item{x}{\code{caracas_symbol}.}

\item{\dots}{further arguments passed to methods}
}
\description{
Math functions
}
