% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{matrix_cross_product}
\alias{matrix_cross_product}
\alias{crossprod_}
\alias{tcrossprod_}
\title{Matrix cross product}
\usage{
crossprod_(x, y = NULL)

tcrossprod_(x, y = NULL)
}
\arguments{
\item{x, y}{caracas matrices}
}
\description{
Matrix cross product
}
\concept{linalg}
