% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\alias{as_piecewise}
\alias{piecewise_cond}
\alias{piecewise_expr}
\title{Create a piecewise object}
\usage{
as_piecewise(pw)

piecewise_cond(pw)

piecewise_expr(pw)
}
\arguments{
\item{pw}{A caracas object containg a Piecewise specification}
}
\value{
A list
}
\description{
Create a piecewise object
}
\examples{
if (has_sympy()) {
  library(caracas)
  def_sym(r, n, j)
  sum1 <- sum_(r^j, var=j, lower=0, upper=n)
  pw <- as_piecewise(sum1)
  pw[[2]]$expr

  pw |> piecewise_cond()
  pw |> piecewise_expr()

}

}
\author{
Søren Højsgaard
}
