% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.caracas_factor_list}
\alias{print.caracas_factor_list}
\title{Print solution}
\usage{
\method{print}{caracas_factor_list}(x, ...)
}
\arguments{
\item{x}{A \code{caracas_factor_list}}

\item{\dots}{Passed to \code{\link[=print.caracas_symbol]{print.caracas_symbol()}}}
}
\description{
Print solution
}
\concept{output}
