% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.caracas_solve_sys_sol}
\alias{print.caracas_solve_sys_sol}
\title{Print solution}
\usage{
\method{print}{caracas_solve_sys_sol}(
  x,
  simplify = getOption("caracas.print.sol.simplify", default = TRUE),
  ...
)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{simplify}{Print solution in a simple format}

\item{\dots}{Passed to \code{\link[=print.caracas_symbol]{print.caracas_symbol()}}}
}
\description{
Print solution
}
\concept{output}
