% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.caracas_symbol}
\alias{print.caracas_symbol}
\title{Print symbol}
\usage{
\method{print}{caracas_symbol}(x, ...)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{\dots}{not used}
}
\description{
Print symbol
}
\concept{output}
