% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.caracas_symbol}
\alias{print.caracas_symbol}
\title{Print symbol}
\usage{
\method{print}{caracas_symbol}(
  x,
  caracas_prefix = TRUE,
  ascii = getOption("caracas.print.ascii", default = FALSE),
  colvec_transposed = getOption("caracas.print.colvec.transposed", default = TRUE),
  ...
)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{caracas_prefix}{Print 'caracas' prefix}

\item{ascii}{\code{TRUE} to print in ASCII format rather than in utf8}

\item{\dots}{not used}
}
\description{
Print symbol
}
\concept{output}
