% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.caracas_symbol}
\alias{print.caracas_symbol}
\title{Print symbol}
\usage{
\method{print}{caracas_symbol}(
  x,
  prompt = getOption("caracas.prompt", default = "c: "),
  method = getOption("caracas.print.method", default = "utf8"),
  rowvec = getOption("caracas.print.rowvec", default = TRUE),
  ...
)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{prompt}{Which prompt/prefix to print (default: 'c: ')}

\item{method}{What way to print (\code{utf8}, \code{prettyascii} or \code{ascii})}

\item{rowvec}{\code{FALSE} to print column vectors as is}

\item{\dots}{not used}
}
\description{
Print symbol
}
\concept{output}
