% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_alg.R
\name{rankMatrix_}
\alias{rankMatrix_}
\title{Rank of matrix}
\usage{
rankMatrix_(x)
}
\arguments{
\item{x}{Numeric or symbolic matrix}
}
\description{
Rank of matrix
}
\examples{
if (has_sympy()) {
  X <- matrix_(paste0("x_",c(1,1,1,1,2,2,2,2,3,4,3,4)), nrow=4)
  X
  rankMatrix_(X)
  colspan(X)
}

}
\concept{linalg}
