% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin-alg.R
\name{reciprocal_matrix}
\alias{reciprocal_matrix}
\title{Elementwise reciprocal matrix}
\usage{
reciprocal_matrix(x, num = 1)
}
\arguments{
\item{x}{Object \code{x}}

\item{num}{The numerator in the result.}
}
\description{
Elementwise reciprocal matrix
}
\concept{linalg}
