% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_lin}
\alias{solve_lin}
\title{Solve a linear system of equations}
\usage{
solve_lin(A, b)
}
\arguments{
\item{A}{matrix}

\item{b}{vector}
}
\description{
Find \code{x} in \code{Ax = b}. If \code{b} not supplied,
the inverse of \code{A} is returned.
}
\examples{
if (has_sympy()) {
  A <- matrix_sym(2, 2, "a")
  b <- vector_sym(2, "b")
  # Inverse of A:
  solve_lin(A) \%*\% A |> simplify()
  # Find x in Ax = b
  x <- solve_lin(A, b)
  A \%*\% x |> simplify()
}

}
\concept{solve}
