% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_sys}
\alias{solve_sys}
\title{Solves a system of non-linear equations}
\usage{
solve_sys(lhs, rhs, vars)
}
\arguments{
\item{lhs}{Equation (or equations as row vector/1xn matrix)}

\item{rhs}{Equation (or equations as row vector/1xn matrix)}

\item{vars}{vector of variable names or symbols}
}
\description{
If called as \code{solve_sys(lhs, vars)}
the roots are found.
If called as \code{solve_sys(lhs, rhs, vars)}
the solutions to \code{lhs = rhs} for \code{vars} are found.
}
\concept{solve}
