% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sh-addition.R
\name{special_matrices}
\alias{special_matrices}
\alias{zeros}
\alias{ones}
\alias{eye}
\title{Special matrices: zeros, ones, eyes}
\usage{
zeros(nrow, ncol)

ones(nrow, ncol)

eye(nrow, ncol)
}
\arguments{
\item{nrow, ncol}{Number of rows and columns of output}
}
\description{
Special matrices: zeros, ones, eyes
}
\examples{

if (has_sympy()){
  zeros(3, 4)
  ones(3, 4)
  eye(3, 4)
}

}
\seealso{
\code{\link[=diag_]{diag_()}}, \code{\link[=matrix_sym]{matrix_sym()}}, \code{\link[=vector_sym]{vector_sym()}}
}
