% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{subs}
\alias{subs}
\title{Substitute af vector of symbols for a vector of values}
\usage{
subs(s, x, v)
}
\arguments{
\item{s}{Expression}

\item{x}{Names of symbol (vector)}

\item{v}{Values for \code{x} (vector)}
}
\description{
Substitute af vector of symbols for a vector of values
}
\examples{
if (has_sympy()) {
   x <- as_sym(paste0('x', 1:3))
   e <- 2*x^2
   e
   subs(e, x, 1:3)
   subs(e, x, x^2)
   
}

}
\seealso{
\code{\link[=subs_single]{subs_single()}}, \code{\link[=subs_lst]{subs_lst()}}
}
\concept{caracas_symbol}
