% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{subs}
\alias{subs}
\title{Substitute symbol for value}
\usage{
subs(s, x, v)
}
\arguments{
\item{s}{Expression}

\item{x}{Name of symbol (character)}

\item{v}{Value for \code{x}}
}
\description{
Substitute symbol for value
}
\examples{
if (has_sympy()) {
   x <- symbol('x')
   e <- 2*x^2
   e
   subs(e, "x", "2")
   y <- as_sym("2")
   subs(e, "x", y)
}

}
