% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{symbol}
\alias{symbol}
\title{Create a symbol}
\usage{
symbol(x)
}
\arguments{
\item{x}{Name to turn into symbol}
}
\value{
A \code{caracas_symbol}
}
\description{
Create a symbol
}
\concept{caracas_symbol}
