% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tex}
\alias{tex}
\title{Export object to TeX}
\usage{
tex(x, zero_as_dot = FALSE, ...)
}
\arguments{
\item{x}{A \code{caracas_symbol}}

\item{zero_as_dot}{Print zero as dots}

\item{\dots}{Other arguments passed along}
}
\description{
Export object to TeX
}
\concept{output}
