% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tex_list}
\alias{tex_list}
\title{Export object to TeX}
\usage{
tex_list(x, zero_as_dot = FALSE, matstr = NULL, ...)
}
\arguments{
\item{x}{A list of \code{caracas_symbol}s and other (simple) R objects (atomics, dataframes, matrices)}

\item{zero_as_dot}{Print zero as dots}

\item{matstr}{Replace \verb{\\begin\{matrix\}} with another environment, e.g. \code{pmatrix}.
If vector of length two, the second element is an optional argument.}

\item{\dots}{Other arguments passed along}
}
\description{
Export object to TeX
}
\examples{
if (has_sympy()) {
  X <- matrix_sym(4,2,"a")
  b <- vector_sym(2,"b")
  y <- vector_sym(4,"y")
  tex_list(list(y, "=", X, b))
}
}
\concept{output}
