% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{vectorfy}
\alias{vectorfy}
\title{Creates symbol vector from list of caracas symbols}
\usage{
vectorfy(x)
}
\arguments{
\item{x}{Symbol to be coerced to vector}
}
\description{
Creates symbol vector from list of caracas symbols
}
